IS-1980: Move oppfolgingsplan into /sider folder
