+++ src/sider/oppfolgingsplan/oppfolgingsplaner/Oppfolgingsplan.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Link } from "react-router-dom";
 import Knapp from "nav-frontend-knapper";
 import { DokumentinfoDTO } from "@/data/oppfolgingsplan/types/DokumentinfoDTO";
-import Feilmelding from "../../Feilmelding";
-import AppSpinner from "../../AppSpinner";
+import Feilmelding from "../../../components/Feilmelding";
+import AppSpinner from "../../../components/AppSpinner";
 import { SYFOOPPFOLGINGSPLANSERVICE_ROOT } from "@/apiConstants";
 import { OppfolgingsplanDTO } from "@/data/oppfolgingsplan/types/OppfolgingsplanDTO";
 import { useDokumentinfoQuery } from "@/data/oppfolgingsplan/oppfolgingsplanQueryHooks";
